Order-Service + extras
- Order Service now automatically assigns the logged in user the order.
- Added Rest Template config to Order-Service.
- You can now only remove orders if the order belongs to you.
- Made price not nullable in Book model.

+++ MicroServices/OrderService/src/main/java/cs4337/group6/OrderService/Services/OrderService.java
@@ -8,7 +8,6 @@ import cs4337.group6.OrderService.Repository.IOrderRepository;
 import cs4337.group6.OrderService.Repository.IUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
 
@@ -24,16 +23,29 @@ public class OrderService {
     @Autowired
     private IUserRepository userRepository;
 
-    public Order CreateOrder(Order order)
+    public Order CreateOrder(Order order, String username) throws Exception
     {
-        User user = userRepository.findById(order.getUserId()).orElseThrow(() -> new RuntimeException("User Not Found"));
+        User user = userRepository.findByUsername(username);
+        if(user == null)
+        {
+            throw new Exception("You are not signed in!");
+        }
         Book book = bookRepository.findById(order.getBookId()).orElseThrow(() -> new RuntimeException("Book Not Found"));
+        order.setUsername(username);
         return orderRepository.save(order);
     }
 
-    public void DeleteOrder(Integer id) {
-        Order orderToDelete = orderRepository.findById(id).orElseThrow(() -> new RuntimeException("Order Not Found"));;
-        orderRepository.delete(orderToDelete);
+    public void DeleteOrder(Integer id, String username) throws Exception {
+        User currentUser = userRepository.findByUsername(username);
+        Order orderToDelete = orderRepository.findById(id).orElseThrow(() -> new RuntimeException("Order Not Found"));
+        if(orderToDelete.getUsername().equals(currentUser.getUsername()))
+        {
+            orderRepository.delete(orderToDelete);
+        }
+        else
+        {
+            throw new Exception("This order does not belong to you!");
+        }
     }
 
     public List<Order> GetAllOrders()
